moved multi tests inside manager tests
